[media] dvb-frontends/stv0367dd: prevent passing zero to intlog10() in QAM_GetSignalToNoise()
In some cases (e.g. when using w_scan on a cable network) it can happen that
the result of the Power*320/regvalavg is zero, which will be passed to
dvb_math's intlog10() function, which in turn will trigger a kernel warning.
This change prevents this by performing the calculation separately and check
for a zero value before passing the value.