[media] dvb-frontends/stv0367: refactor STV0367 PLL Setup
This moves the PLL Setup logic into an own function and adds a configurable
IC speed (53.125 or 58 MHz). This is required for ddbridge which runs its
stv0367 variant at 58 MHz. Also, PLL is now set up for -T and -C aswell.

This commit also updates all users of the stv0367 frontend driver.