[media] dvb-frontends/stv0367: make full reinit on set_frontend() optional
Every time dvb_frontend_ops.set_frontend() is called, an almost full reinit of
the demodulator will be performed. While this might cause a slight delay when
switching channels due to all involved tables being rewritten, it can even be
dangerous in certain causes in that the demod may lock up and requires to be
powercycled (this can happen on Digital Devices hardware). So this adds a
flag if it should be done, and to not change behaviour with existing card
support, it'll be enabled in all cases.

Signed-off-by: Daniel Scheller <d.scheller@gmx.net>