[media] dvb-frontends/stv0367: support reading if_khz from tuner config
Currently, if_khz is set and provided using the configuration var in
struct stv0367_config. However, in some constellations, the value might be
different for differing channel bandwidths or even -T and -C operation. When
e.g. used in conjunction with TDA18212 tuners, the tuner frontend might be
aware of the different freqs. This factors if_khz retrieval in a function,
which checks a new flag if an automatic retrieval attempt should be made, and
if the tuner provides it, use it whenever needed.

Signed-off-by: Daniel Scheller <d.scheller@gmx.net>