[media] dvb-frontends/cxd2841er: sleep on delivery system switch
Again, maybe depending on the hardware constellation (demod/tuner combo),
the demodulator might be needed to put to sleep when the requested
delivery system differs from the last used one (this is for example the
case when tuning from a DVB-T to a DVB-T2 mux or vice versa), else
proper tuning and TS delivery will fail for the "new" delivery system.
This is the case at least with Digital Devices hardware, so add a
flag to toggle this behaviour.

Signed-off-by: Daniel Scheller <d.scheller@gmx.net>