[media] dvb-frontends/cxd2841er: cached TSCONFIG register
A strange behaviour was discovered in that after initialising the demod and
later on calling set_ts_clock_mode(), the TS config register was flipped
around in a way it isn't feasible anymore to just toggle some bits in it.
Since right after init everything is right, instead of setting up some cfg
bits, read out the value, set up all required bits for the wanted TS config
and then (when needed) poke this exact value. This will prevent from random
failures due to the TS configuration not being correct.

Signed-off-by: Daniel Scheller <d.scheller@gmx.net>