[media] dvb-frontends/cxd2841er: prevent multiple inits for HW_DDB
On DD hardware constellation, the combination of putting the device back
to STATE_SHUTDOWN_TC, setting/disabling the osc and demod mode and later
on calling init_tc() again will cause random things to be not properly
being set up afterwards, causing at least the TS mode to be wrong. So
don't set this mode when HW_DDB is set and also NULL the init fe_op
and perform init from _attach() afterwards.

Signed-off-by: Daniel Scheller <d.scheller@gmx.net>