media: dvb-frontends/stv0910: make various local variables static
The list stvlist and arrays padc_lookup, s1_sn_lookup and s2_sn_lookup
do not need to be in global scope, so make them all static.

Cleans up a bunch of smatch warnings:
symbol 'padc_lookup' was not declared. Should it be static?
symbol 's1_sn_lookup' was not declared. Should it be static?
symbol 's2_sn_lookup' was not declared. Should it be static?
symbol 'stvlist' was not declared. Should it be static?

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Daniel Scheller <d.scheller@gmx.net>